Allow specifying `serviceAccountName` for Terrakube components
In AWS EKS, Service Accounts can be used to securely grant Pods access
to AWS API using feature [IAM roles for service accounts][1].

Allow user to specify Service Account name for Terrakube components to
be able to use this feature.

In the future, if deemed desirable, the chart could also create a
Service Account by itself, but for now I decided to keep it out of
scope.

[1]: https://docs.aws.amazon.com/eks/latest/userguide/iam-roles-for-service-accounts.html